+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Next
 
+- [New Model] Add waterbox support for Roborock S7 Max Ultra. ([#TBD](https://github.com/homebridge-xiaomi-roborock-vacuum/homebridge-xiaomi-roborock-vacuum/pull/TBD))
+
 ## 0.33.0
 
-- [New Model] Add waterbox support for Roborock Q8 Max+ Thank you @jmago17 ([#1026](https://github.com/homebridge-xiaomi-roborock-vacuum/homebridge-xiaomi-roborock-vacuum/pull/1026))
+- [New Model] Add waterbox support for Roborock Q8 Max+. Thank you, @jmago17 ([#1026](https://github.com/homebridge-xiaomi-roborock-vacuum/homebridge-xiaomi-roborock-vacuum/pull/1026))
 
 ## 0.32.1